compute: one ComputeSession, and the rail that can open one says so in the type
`ComputeSession` was two structs sharing a name — eleven fields on the Metal arm
and one on the Vulkan arm — with `open`, `encode_control`, `encode_icb` and
`finish` each a `cfg`-else inside one body, the Vulkan half of every one a
`NoMetal` refusal. Two shapes under one name is how a type starts disagreeing
with itself across a feature boundary, and it is why `execute_dispatch_nested`
could not become a `Backend` method in the previous commit.

* `ComputeSession` is one struct on every arm, holding a `SessionRail`.
  `compute_session::metal` owns the encoder, the command buffer, the five
  Objective-C retain lists, the deferred nested-dispatch jobs, and the ICB
  parent-encoder inheritance that reads all of them.
* `SessionRail` has one variant per rail that can hold something open. Vulkan is
  not one — `engine::execute_compute_request` is one-shot per dispatch — so it
  contributes none and refuses at `Backend::open_compute_session`. The
  `Unbacked` variant carries an `Infallible`, so it is not a state a session can
  be in; it exists only because a build with no session-capable rail would
  otherwise make the enum empty, and rustc treats a `&mut` to an uninhabited
  type as something every method still has to match on.
* `Backend` gains `open_compute_session` and `execute_dispatch_nested`. The
  capability is now refused once, where it is asked for, instead of at four
  entry points a session would have had to decline one at a time.
* `control_depth` moves into the rail that moves it, rather than sitting beside
  it as a second copy of the same counter.

Guest-visible refusals are unchanged. `compute_session_no_vulkan_path` and
`compute_nested_no_vulkan_path` keep their slugs and their call sites; the three
`no_session_rail` slugs replace refusals that were already unreachable, because
a rail that cannot open a session never had one to encode onto.

One Vulkan test changes shape: it built a `ComputeSession` by hand to assert
that a nested dispatch refuses. There is no such value to build now, so it
asserts the same contract where the contract is decided — `open_compute_session`
refuses with the slug it always did.

No behavior change on either arm. Metal 1315 pass, Vulkan 2173 pass, both
serial; feature-matrix green with per-arm counts held.
