+++ crates/reims-vgpu/src/runtime/icb/tests.rs
@@ -979,7 +979,7 @@ fn materialize_and_execute_empty_range() {
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn fill_and_execute_mul3add1_writeback() {
-    use crate::runtime::compute_session::ComputeSession;
+    use crate::backend::Backend as _;
 
     let (_guard, mtlb, mut host, mut state) = mul3add1_fixture();
 
@@ -1028,7 +1028,9 @@ fn fill_and_execute_mul3add1_writeback() {
     assert_eq!(icb_a.size(), 1);
 
     // Product execute 0xe4 range [0,1] on a compute session + writeback.
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(
@@ -1920,7 +1922,7 @@ fn decode_encode_command_slot_roundtrip() {
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn buffer_backed_fill_execute_mul3add1() {
-    use crate::runtime::compute_session::ComputeSession;
+    use crate::backend::Backend as _;
 
     let (_guard, mtlb, mut host, mut state) = mul3add1_fixture();
 
@@ -1979,7 +1981,9 @@ fn buffer_backed_fill_execute_mul3add1() {
     assert_eq!(mem.gva, cmd_mem_gva);
     assert_eq!(mem.byte_len, layout.command_size as u64);
 
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(
@@ -4057,7 +4061,7 @@ fn decode_encode_barrier_and_threadgroup_memory() {
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn fill_compute_barrier_and_tg_memory_execute() {
-    use crate::runtime::compute_session::ComputeSession;
+    use crate::backend::Backend as _;
 
     let (_guard, mtlb, mut host, mut state) = mul3add1_fixture();
 
@@ -4096,7 +4100,9 @@ fn fill_compute_barrier_and_tg_memory_execute() {
     )
     .expect("fill with barrier+tg");
 
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(
@@ -4122,8 +4128,8 @@ fn fill_compute_barrier_and_tg_memory_execute() {
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn inherit_buffers_encoder_kernel_mul3add1() {
+    use crate::backend::Backend as _;
     use crate::runtime::compute_exec::{ComputeAccum, ComputeBufferBind};
-    use crate::runtime::compute_session::ComputeSession;
 
     let (_guard, mtlb, mut host, mut state) = mul3add1_fixture();
 
@@ -4171,7 +4177,9 @@ fn inherit_buffers_encoder_kernel_mul3add1() {
         has_attribute_stride: false,
     });
 
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(&mut state, &mut host, 1, &cmd, &acc),
@@ -4196,8 +4204,8 @@ fn inherit_buffers_encoder_kernel_mul3add1() {
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn inherit_pipeline_encoder_kernel_mul3add1() {
+    use crate::backend::Backend as _;
     use crate::runtime::compute_exec::ComputeAccum;
-    use crate::runtime::compute_session::ComputeSession;
     use crate::runtime::decode::resource::ICB_FLAG_INHERIT_PIPELINE_STATE;
 
     let (_guard, mtlb, mut host, mut state) = mul3add1_fixture();
@@ -4238,7 +4246,9 @@ fn inherit_pipeline_encoder_kernel_mul3add1() {
     let mut acc = ComputeAccum::default();
     acc.set_pipeline(6);
 
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(&mut state, &mut host, 1, &cmd, &acc),
@@ -4335,9 +4345,9 @@ fn put_type7_sampler(host: &mut FakeHost, state: &DeviceState, obj_ref: u32, nor
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn icb_parent_encoder_texture_and_sampler_binds() {
+    use crate::backend::Backend as _;
     use crate::contract::pixel_format::MTL_FORMAT_RGBA8_UNORM;
     use crate::runtime::compute_exec::{ComputeAccum, ComputeSamplerBind, ComputeTextureBind};
-    use crate::runtime::compute_session::ComputeSession;
 
     let (_guard, mtlb, mut host, mut state) = mul3add1_fixture();
 
@@ -4403,7 +4413,9 @@ fn icb_parent_encoder_texture_and_sampler_binds() {
         has_lod_clamp: false,
     });
 
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(&mut state, &mut host, 1, &cmd, &acc),
@@ -4446,9 +4458,9 @@ fn icb_parent_encoder_texture_and_sampler_binds() {
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn icb_argument_buffer_storage_texture_xyplane() {
+    use crate::backend::Backend as _;
     use crate::contract::pixel_format::MTL_FORMAT_RGBA8_UINT;
     use crate::runtime::compute_exec::{ComputeAccum, ComputeTextureBind};
-    use crate::runtime::compute_session::ComputeSession;
 
     let _guard = icb_test_guard();
     let mtlb = std::fs::read(
@@ -4505,7 +4517,9 @@ fn icb_argument_buffer_storage_texture_xyplane() {
         texture_ref: 11,
     });
 
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(&mut state, &mut host, 1, &cmd, &acc),
@@ -4543,9 +4557,9 @@ fn icb_argument_buffer_storage_texture_xyplane() {
 #[test]
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
 fn icb_argument_buffer_sample_and_write() {
+    use crate::backend::Backend as _;
     use crate::contract::pixel_format::{MTL_FORMAT_RGBA8_UINT, MTL_FORMAT_RGBA8_UNORM};
     use crate::runtime::compute_exec::{ComputeAccum, ComputeSamplerBind, ComputeTextureBind};
-    use crate::runtime::compute_session::ComputeSession;
 
     let _guard = icb_test_guard();
     let mtlb = std::fs::read(
@@ -4633,7 +4647,9 @@ fn icb_argument_buffer_sample_and_write() {
         has_lod_clamp: false,
     });
 
-    let mut session = ComputeSession::open(0).expect("session");
+    let mut session = crate::backend::selected()
+        .open_compute_session(0)
+        .expect("session");
     let cmd = execute_icb_command(9, 0, 1);
     assert_eq!(
         session.encode_icb(&mut state, &mut host, 1, &cmd, &acc),
